+++ src/test/java/online/strongnation/unit/entity/CountryDataSavingTest.java
@@ -1,15 +1,16 @@
 package online.strongnation.unit.entity;
 
+import online.strongnation.exception.CountryNotFoundException;
 import online.strongnation.model.dto.CategoryDTO;
 import online.strongnation.model.entity.Category;
 import online.strongnation.model.entity.Country;
 import online.strongnation.model.entity.CountryCategory;
-import online.strongnation.exception.CountryNotFoundException;
 import online.strongnation.repository.CountryRepository;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.orm.jpa.DataJpaTest;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 import static org.assertj.core.api.AssertionsForClassTypes.assertThat;
@@ -29,7 +30,7 @@ public class CountryDataSavingTest {
         country.setName(name);
         CountryCategory regionCategory = new CountryCategory();
         Category category = new Category();
-        category.setNumber(9.f);
+        category.setNumber(BigDecimal.valueOf(9.f));
         category.setName("food");
         category.setUnits("kg");
         regionCategory.setCategory(category);
